+++ lib/aixm/constants.rb
@@ -5,4 +5,6 @@ module AIXM
   GROUND = AIXM::Z.new(alt: 0, code: :QFE).freeze
   UNLIMITED = AIXM::Z.new(alt: 999, code: :QNE).freeze
 
+  H24 = AIXM::Schedule.new(code: :H24).freeze
+
 end